zebra: include our own copies of some linux kernel headers
This is the same strategy adopted by the iproute2 project to deal with
building issues on old systems. These header files don't need to be in
sync with upstream, they only need to be updated when necessary (e.g. if
we want to use a new feature introduced by a recent kernel).

Fixes #962 using the solution suggested by David Lamparter.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>